Security hardening: close critical money-safety and secret-leak holes
C1: withdraw double-pay TOCTOU — BroadcastUncertainError with pre-computed
    tx_hash; NEVER auto-refund after broadcast attempt.
C2: x402 redirect/drain — dedicated X402_RECEIVE_ADDRESS, hot-wallet
    deposit rejection, receive-address verification.
C3: onchain_market.py broken imports — rpc_url, USDC_ADDRESS, send_eth.
C4: outcome.py config.HOT_WALLET AttributeError — module-level _HOT_WALLET.
C5: MCP server unauthenticated on 0.0.0.0 — auth middleware + localhost.
H1: XSS in me.html — esc() function, all innerHTML escaped.
H2: Session cookie Secure behind proxy — x-forwarded-proto check.
H3: config.validate() at import time of server.py.
H7: /qr OOM — size bounded 64..1024.
CI: conftest.py pytest.exit, compose exit-code, lint covers agent/scripts,
    valid BOT_TOKEN, SECRET_KEY in test env.
Lint: ruff clean across bot web tests deploy agent scripts.

589 tests passed, 0 failed.

+++ agent/signals.py
@@ -11,8 +11,10 @@ from the SURGE-AGENT-PROMPT.
 """
 
 import time
-from bot.ledger import ledger
+
 from bot import config
+from bot.ledger import ledger
+
 from . import caps
 
 
@@ -34,7 +36,7 @@ async def sell_signal(
         title = f"Signal: Market #{market_id} analysis ({time.strftime('%Y-%m-%d %H:%M')})"
 
     title = title[: config.PAYWALL_MAX_TITLE_LEN]
-    price_micro = int(round(price_usdc * 10**config.USDC_DECIMALS))
+    price_micro = round(price_usdc * 10**config.USDC_DECIMALS)
     tg_id = 0  # agent's tg_id (0 = system)
 
     content = (
